move InteractiveScriptStateListener to tools project. rename onScriptComplete() to onScriptSuccess() use a working pair of scriptSig and scriptPubKey in the example.
