+++ tools/src/main/java/org/bitcoinj/tools/InteractiveScriptStateListener.java
@@ -1,7 +1,11 @@
-package org.bitcoinj.script;
+package org.bitcoinj.tools;
 
 import org.bitcoinj.core.*;
 import org.bitcoinj.params.MainNetParams;
+import org.bitcoinj.script.Script;
+import org.bitcoinj.script.ScriptChunk;
+import org.bitcoinj.script.ScriptOpCodes;
+import org.bitcoinj.script.ScriptStateListener;
 
 import java.util.*;
 import java.util.regex.Matcher;
@@ -21,14 +25,22 @@ public class InteractiveScriptStateListener extends ScriptStateListener {
 
     public static void main(String[] args) {
 
+        /**
+         *
+         */
+
         NetworkParameters params = MainNetParams.get();
 
-        String rawTransaction = "0100000002d8c8df6a6fdd2addaf589a83d860f18b44872d13ee6ec3526b2b470d42a96d4d000000008b483045022100b31557e47191936cb14e013fb421b1860b5e4fd5d2bc5ec1938f4ffb1651dc8902202661c2920771fd29dd91cd4100cefb971269836da4914d970d333861819265ba014104c54f8ea9507f31a05ae325616e3024bd9878cb0a5dff780444002d731577be4e2e69c663ff2da922902a4454841aa1754c1b6292ad7d317150308d8cce0ad7abffffffff2ab3fa4f68a512266134085d3260b94d3b6cfd351450cff021c045a69ba120b2000000008b4830450220230110bc99ef311f1f8bda9d0d968bfe5dfa4af171adbef9ef71678d658823bf022100f956d4fcfa0995a578d84e7e913f9bb1cf5b5be1440bcede07bce9cd5b38115d014104c6ec27cffce0823c3fecb162dbd576c88dd7cda0b7b32b0961188a392b488c94ca174d833ee6a9b71c0996620ae71e799fc7c77901db147fa7d97732e49c8226ffffffff02c0175302000000001976a914a3d89c53bb956f08917b44d113c6b2bcbe0c29b788acc01c3d09000000001976a91408338e1d5e26db3fce21b011795b1c3c8a5a5d0788ac00000000";
+        String rawTx1 = "010000000193e3073ecc1d27f17e3d287ccefdfdba5f7d8c160242dbcd547b18baef12f9b31a0000006b483045022100af501dc9ef2907247d28a5169b8362ca494e1993f833928b77264e604329eec40220313594f38f97c255bcea6d5a4a68e920508ef93fd788bcf5b0ad2fa5d34940180121034bb555cc39ba30561793cf39a35c403fe8cf4a89403b02b51e058960520bd1e3ffffffff02b3bb0200000000001976a914f7d52018971f4ab9b56f0036958f84ae0325ccdc88ac98100700000000001976a914f230f0a16a98433eca0fa70487b85fb83f7b61cd88ac00000000";
+        String rawtx2 = "01000000011c9c6bb48a760cf656480a33340331859185b336f1effa30c8f76a8872cf27ed000000006a47304402201c999cf44dc6576783c0f55b8ff836a1e22db87ed67dc3c39515a6676cfb58e902200b4a925f9c8d6895beed841db135051f8664ab349f2e3ea9f8523a6f47f93883012102e58d7b931b5d43780fda0abc50cfd568fcc26fb7da6a71591a43ac8e0738b9a4ffffffff029b010100000000001976a9140f0fcdf818c0c88df6860c85c9cc248b9f37eaff88ac95300100000000001976a9140663d2403f560f8d053a25fbea618eb47071617688ac00000000";
 
-        byte[] txBytes = HEX.decode(rawTransaction);
-        Transaction tx = new Transaction(params, txBytes);
-        Script scriptSig = tx.getInput(0).getScriptSig();
-        Script scriptPubKey = tx.getOutput(0).getScriptPubKey();
+
+        byte[] tx1Bytes = HEX.decode(rawTx1);
+        Transaction tx1 = new Transaction(params, tx1Bytes);
+        byte[] tx2Bytes = HEX.decode(rawtx2);
+        Transaction tx2 = new Transaction(params, tx2Bytes);
+        Script scriptPubKey = tx1.getOutput(0).getScriptPubKey();
+        Script scriptSig = tx2.getInput(0).getScriptSig();
 
         LinkedList<byte[]> stack = new LinkedList();
 
@@ -40,7 +52,7 @@ public class InteractiveScriptStateListener extends ScriptStateListener {
 
         System.out.println("\n***Executing scriptPubKey***\n");
         script = scriptPubKey;
-        Script.executeDebugScript(null, 0, script, stack, Coin.ZERO, Script.ALL_VERIFY_FLAGS, listener);
+        Script.executeDebugScript(tx1, 0, script, stack, Coin.ZERO, Script.ALL_VERIFY_FLAGS, listener);
 
 
     }
@@ -81,12 +93,16 @@ public class InteractiveScriptStateListener extends ScriptStateListener {
         Collections.reverse(reverseStack);
         System.out.println("Stack:");
 
-        for (byte[] bytes: reverseStack) {
-            System.out.println(HEX.encode(bytes));
+        if (reverseStack.isEmpty()) {
+          System.out.println("empty");
+        } else {
+            for (byte[] bytes : reverseStack) {
+                System.out.println(HEX.encode(bytes));
+            }
         }
         System.out.println();
 
-        if (getAltstack().size() > 0) {
+        if (!getAltstack().isEmpty()) {
             reverseStack = new ArrayList<byte[]>(getAltstack());
             Collections.reverse(reverseStack);
             System.out.println("Alt Stack:");
@@ -97,7 +113,7 @@ public class InteractiveScriptStateListener extends ScriptStateListener {
             System.out.println();
         }
 
-        if (getIfStack().size() > 0) {
+        if (!getIfStack().isEmpty()) {
             List<Boolean>reverseIfStack = new ArrayList<Boolean>(getIfStack());
             Collections.reverse(reverseIfStack);
             System.out.println("If Stack:");
@@ -122,8 +138,8 @@ public class InteractiveScriptStateListener extends ScriptStateListener {
     }
 
     @Override
-    public void onScriptComplete() {
-
+    public void onScriptSuccess() {
+        System.out.println("Script success.");
     }
 
     private String truncateData(String scriptString) {
